Cambios

+++ paseoModernoBk/src/main/java/com/example/paseomodernobk/Service/CategoryService.java
@@ -1,9 +1,13 @@
 package com.example.paseomodernobk.Service;
 
 import com.example.paseomodernobk.Entity.CategoryEntity;
+import com.example.paseomodernobk.Entity.ProductEntity;
 import com.example.paseomodernobk.Exceptions.ResourceNotFoundException;
 import com.example.paseomodernobk.Repository.CategoryRepository;
+import com.example.paseomodernobk.Repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,8 +19,11 @@ public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
 
-    public List<CategoryEntity> getAllCategories() {
-        return categoryRepository.findAll();
+    @Autowired
+    private ProductRepository productRepository;
+
+    public Page<CategoryEntity> getAllCategories(Pageable pageable) {
+        return categoryRepository.findAll(pageable);
     }
 
     public Optional<CategoryEntity> getCategoryById(Long id) {
@@ -38,7 +45,19 @@ public class CategoryService {
     }
 
     public void deleteCategoryById(Long id) {
-        categoryRepository.deleteById(id);
+        final Long idSinCategoria = Long.valueOf(1);
+        CategoryEntity category = categoryRepository.findById(idSinCategoria).orElse(null);
+
+        if (category != null) {
+            List<ProductEntity> productList = productRepository.findAllByCategory(id);
+
+            productList.forEach(productEntity -> {
+                productEntity.setCategory(category);
+            });
+            productRepository.saveAll(productList);
+            categoryRepository.deleteById(id);
+        }
     }
+
 }
